#1704 Moved action generation from tg-app-template to tg-app-menu to overcaome the need of  slot creation which should fix the Safari slot problem described in this issue
