+++ platform-web-ui/src/main/java/ua/com/fielden/platform/web/menu/impl/MainMenuBuilder.java
@@ -49,7 +49,6 @@ public class MainMenuBuilder implements IMainMenuBuilderWithLayout {
     public Menu getMenu() {
         return new Menu().
                 setMenu(mainMenu.getModules()).
-                setActions(mainMenu.generateMenuActions()).
                 setMinCellWidth(tileLayout.getMinCellWidth() + "px").
                 setMinCellHeight(tileLayout.getMinCellHeight() + "px").
                 setWhenDesktop(tileLayout.getLayout(Device.DESKTOP, null).get()).
