+++ platform-web-ui/src/main/java/ua/com/fielden/platform/web/menu/module/impl/WebMenuModule.java
@@ -2,9 +2,13 @@ package ua.com.fielden.platform.web.menu.module.impl;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
+import ua.com.fielden.platform.menu.Action;
 import ua.com.fielden.platform.menu.Module;
 import ua.com.fielden.platform.web.centre.api.actions.EntityActionConfig;
+import ua.com.fielden.platform.web.centre.api.resultset.impl.FunctionalActionElement;
+import ua.com.fielden.platform.web.centre.api.resultset.impl.FunctionalActionKind;
 
 public class WebMenuModule {
 
@@ -67,14 +71,15 @@ public class WebMenuModule {
         return actions;
     }
 
-    public Module getModule() {
+    public Module getModule(final int startActionIndex) {
         final Module module = new Module().
                 setBgColor(bgColor).
                 setCaptionBgColor(captionBgColor).
                 setIcon(icon).
                 setDetailIcon(detailIcon).
                 setKey(title).
-                setDesc(description);
+                setDesc(description).
+                setActions(generateMenuActions(startActionIndex));
         //TODO module menu can not be null. Right now platform supports modules with view. This case should be covered with separate issue.
         if (this.menu != null) {
             module.setMenu(menu.getMenu());
@@ -82,5 +87,39 @@ public class WebMenuModule {
             module.setView(view.getView());
         }
         return module;
+    }
+
+    public List<Action> generateMenuActions(final int startActionIndex) {
+        int numberOfActions = startActionIndex;
+        final List<Action> actions = new ArrayList<>();
+        for (final EntityActionConfig config : getActions()) {
+            final FunctionalActionElement actionElement = new FunctionalActionElement(config, numberOfActions++, FunctionalActionKind.TOP_LEVEL);
+            final Map<String, Object> attributes = actionElement.createAttributes();
+            final Action action = new Action();
+            action.setKey(attributes.get("element-name").toString());
+            action.setDesc(attributes.get("short-desc").toString());
+            action.setUiRole(attributes.get("ui-role").toString());
+            action.setLongDesc(attributes.get("long-desc").toString());
+            action.setShortcut(attributes.getOrDefault("shortcut", "").toString());
+            action.setIcon(attributes.get("icon").toString());
+            action.setIconStyle(attributes.get("icon-style").toString());
+            action.setRefreshParentCentreAfterSave((Boolean)attributes.get("should-refresh-parent-centre-after-save"));
+            action.setComponentUri(attributes.get("component-uri").toString());
+            action.setDynamicAction((Boolean)attributes.getOrDefault("dynamic-action", false));
+            action.setNumberOfAction((Integer)attributes.get("number-of-action"));
+            action.setActionKind(attributes.get("action-kind").toString());
+            action.setElementAlias(attributes.getOrDefault("element-alias", "").toString());
+            action.setChosenProperty(attributes.getOrDefault("chosen-property", "").toString());
+            action.setRequireSelectionCriteria(attributes.get("require-selection-criteria").toString());
+            action.setRequireSelectedEntities(attributes.get("require-selected-entities").toString());
+            action.setRequireMasterEntity(attributes.get("require-master-entity").toString());
+            action.setModuleName(title);
+            action.setPreAction(actionElement.createPreAction());
+            action.setPostActionSuccess(actionElement.createPostActionSuccess());
+            action.setPostActionError(actionElement.createPostActionError());
+            action.setAttrs(actionElement.createElementAttributes(true));
+            actions.add(action);
+        }
+        return actions;
     }
 }
