+++ platform-web-ui/src/main/web/ua/com/fielden/platform/web/views/tg-app-menu.js
@@ -117,7 +117,31 @@ const template = html`
                                 </div>
                             </div>
                             <div class="tile-toolbar layout horizontal wrap" action-disabled$="[[_isDisabled(item)]]">
-                                <slot name$="[[item.key]]"></slot>
+                                <template is="dom-repeat" items="[[item.actions]]" as="action">
+                                    <tg-ui-action slot$="[[action.moduleName]]" 
+                                        show-dialog="[[action._showDialog]]"
+                                        toaster="[[action.toaster]]"
+                                        short-desc="[[action.desc]]" 
+                                        long-desc="[[action.longDesc]]"
+                                        icon="[[action.icon]]"
+                                        component-uri="[[action.componentUri]]"
+                                        element-name="[[action.key]]"
+                                        action-kind="[[action.actionKind]]" 
+                                        number-of-action="[[action.numberOfAction]]"
+                                        dynamic-action="[[action.dynamicAction]]"
+                                        attrs="[[action.attrs]]"
+                                        create-context-holder="[[action._createContextHolder]]" 
+                                        require-selection-criteria="[[action.requireSelectionCriteria]]" 
+                                        require-selected-entities="[[action.requireSelectedEntities]]" 
+                                        require-master-entity="[[action.requireMasterEntity]]"
+                                        pre-action="[[action.preAction]]"
+                                        post-action-success="[[action.postActionSuccess]]" 
+                                        post-action-error="[[action.postActionError]]" 
+                                        should-refresh-parent-centre-after-save="[[action.refreshParentCentreAfterSave]]"
+                                        ui-role="[[action.uiRole]]"
+                                        icon-style="[[action.iconStyle]]">
+                                    </tg-ui-action>
+                                </template>
                             </div>
                         </div>
